move `completion` and `gotodef` packages to `search` package
